requirements.txt
Django==1.8.6
djangorestframework==3.3.1
djrill==1.4.0
requests==2.8.1
wheel==0.24.0
